requirements.txt
requests
pillow
torch
numpy
